Move subcmds to a new folder
