+++ cmd/analyzer/subcmds/parse_args.go
@@ -4,7 +4,7 @@ Copyright 2023- IBM Inc. All Rights Reserved.
 SPDX-License-Identifier: Apache-2.0
 */
 
-package main
+package subcmds
 
 import "github.com/np-guard/vpc-network-config-analyzer/pkg/vpcmodel"
