+++ cmd/analyzer/subcmds/root.go
@@ -4,7 +4,7 @@ Copyright 2023- IBM Inc. All Rights Reserved.
 SPDX-License-Identifier: Apache-2.0
 */
 
-package main
+package subcmds
 
 import (
 	"strings"
